Add data and code for YTD Border Patrol Encounters
The dataset already in github is for historic data, which stops at the current fiscal year (technically includes first two months of fiscal year).

There is a separate dashboard for YTD encounter data: https://www.cbp.gov/newsroom/stats/southwest-land-border-encounters

It is hard to scrape: https://stackoverflow.com/questions/79784978/importing-a-table-from-a-webpage-as-a-dataframe-in-python/79785744

But has links to the csv files which power it: https://www.cbp.gov/newsroom/stats/cbp-public-data-portal

Which links to: https://www.cbp.gov/document/stats/southwest-land-border-encounters

I got the latest data ("FY22 - FY25 (FYTD) Southwest Land Border Encounters - August").

The notebook checked in ingests, cleans and graphs that data. And then reformats it so that it can be merged with the historic dataset, and then graphs the result.