Refactor: move source code into its own directory
Apparently pytest works best in this structure. Note that there's now __init__.py file in immigration_enforcement/immigration_enforcement. This is apparently helpful / necessary for pytest as well.

pytest now works. Note that I changed the import in the file to "from immigration_enforcement.detentions import get_col_prefix". That immigration_enforcement.detentions works, I think, because immigration_enforcement is now a package.

Add a ci test for pytest

excludes mypy from tests. That just seems like overkill and not particularly helpful.

TODO: It looks like in this new structure, it's common for people to add a /notebooks directory (censusdis does). The core of this package are two modules that ingest and analyze two different datasets. I think I have a notebook for one next to the module, and used to have a notebook for another one. I should put the notebook documentation there.