Add nbval to dev dependencies.
This allows me to run pytest on the existing notebooks. Right now I just care that they work (i.e. execute without failure). That that I'm committing in a failing test (just type "pytest --nbval notebooks/"). This is because of a path issue - the `encounters` module is no longer in the same directory as the notebook.