Making caching of detentions module data optional
The data, which requires an API call, should always be cached when accessed via the app. But in notebooks, it should never be cached. The reason is that the caching mechanism uses Streamlit, which requires the context of the app in order to work. And so calling it from the notebook generates a lot of warnings.